+++ core/core-auth/src/test/java/com/barlow/core/auth/authentication/token/AccessTokenProviderTest.java
@@ -6,8 +6,7 @@ import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.barlow.core.auth.authentication.core.MemberPrincipal;
 import com.barlow.core.auth.config.TestKeyConfig;
-import com.barlow.core.support.crypto.AlgorithmConfig;
-import com.barlow.core.support.jwt.TokenConfig;
+import com.barlow.core.auth.config.TokenConfig;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
